+++ homeworks/vladyslav.kovalov_kovalov/7-oop/src/app.ts
@@ -1,0 +1,9 @@
+import { Gallery } from './gallery';
+import { Uploader } from './uploader';
+import { FullScreen } from './fullscreen';
+
+new Gallery(
+    document.querySelector('[data-gallery]') as HTMLElement,
+    FullScreen,
+    Uploader
+);
