+++ homeworks/vladyslav.kovalov_kovalov/7-oop/webpack.config.js
@@ -4,7 +4,7 @@ module.exports = {
 	mode: 'development',
 	devtool: 'inline-source-map',
 	entry: {
-		app: './src/index.ts',
+		app: './src/app.ts',
 	},
 	module: {
     rules: [
